python-generators-0x00/requirements.txt
python-decouple==3.8

